vscode-ext: overwrite with caesar-ext
