--- caesar-ext/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Change Log
-
-All notable changes to the "caesar" extension will be documented in this file.
-
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
-
-## [Unreleased]
-
-- Initial release
