+++ vscode-ext/.vscodeignore
@@ -1,4 +1,11 @@
 .vscode/**
 .vscode-test/**
+src/**
 .gitignore
+.yarnrc
 vsc-extension-quickstart.md
+**/tsconfig.json
+**/.eslintrc.json
+**/*.map
+**/*.ts
+**/.vscode-test.*
