+++ vscode-ext/images/verified.png
