+++ vscode-ext/package.json
@@ -1,21 +1,32 @@
 {
-  "name": "heyvl",
-  "displayName": "HeyVL",
-  "description": "HeyVL language support",
+  "name": "caesar",
+  "displayName": "caesar",
+  "description": "VSCode Extension of the Deductive Probabilistic Program Verifier Caesar",
   "version": "0.0.1",
   "engines": {
-    "vscode": "^1.71.0"
+    "vscode": "^1.87.0"
   },
   "categories": [
-    "Programming Languages"
+    "Other"
   ],
-  "activationEvents": [],
   "main": "./out/extension.js",
   "contributes": {
     "commands": [
       {
-        "command": "helloworld.helloWorld",
-        "title": "Hello World"
+        "command": "caesar.restartServer",
+        "title": "Caesar: Restart Server"
+      },
+      {
+        "command": "caesar.startServer",
+        "title": "Caesar: Start Server"
+      },
+      {
+        "command": "caesar.stopServer",
+        "title": "Caesar: Stop Server"
+      },
+      {
+        "command": "caesar.verify",
+        "title": "Caesar: Verify"
       }
     ],
     "languages": [
@@ -37,17 +48,78 @@
         "scopeName": "source.heyvl",
         "path": "./syntaxes/heyvl.tmLanguage.json"
       }
+    ],
+    "configuration": {
+      "type": "object",
+      "title": "Caesar",
+      "properties": {
+        "caesar.uI.gutterIcons.showGutterIcons": {
+          "type": "boolean",
+          "default": true,
+          "description": "Specifies whether to show icons on the gutter about the verification status."
+        },
+        "caesar.uI.inlineGhostText.showInlineGhostText": {
+          "type": "boolean",
+          "default": true,
+          "description": "Specifies whether to show inline ghost text about the wp/wlp/ert."
+        },
+        "caesar.uI.statusBar.showStatusBar": {
+          "type": "boolean",
+          "default": true,
+          "description": "Specifies whether to show the status bar."
+        },
+        "caesar.automaticVerification": {
+          "type": "string",
+          "default": "onsave",
+          "enum": [
+            "never",
+            "onsave"
+          ],
+          "description": "Specifies when Caesar automatically verifies the file."
+        }
+      }
+    },
+    "colors": [
+      {
+        "id": "caesar.inlineGhostBackgroundColor",
+        "description": "Specifies the background color of the inline ghost texts ",
+        "defaults": {
+          "dark": "#00000000",
+          "light": "#00000000",
+          "highContrast": "#00000000"
+        }
+      },
+      {
+        "id": "caesar.inlineGhostForegroundColor",
+        "description": "Specifies the foreground color of the inline ghost texts",
+        "defaults": {
+          "dark": "#99999959",
+          "light": "#99999959",
+          "highContrast": "#99999999"
+        }
+      }
     ]
   },
   "scripts": {
     "vscode:prepublish": "npm run compile",
     "compile": "tsc -p ./",
-    "watch": "tsc -watch -p ./"
+    "watch": "tsc -watch -p ./",
+    "pretest": "npm run compile && npm run lint",
+    "lint": "eslint src --ext ts",
+    "test": "vscode-test"
   },
   "devDependencies": {
-    "@types/node": "^8.10.25",
-    "@types/vscode": "^1.71.0",
-    "tslint": "^5.16.0",
-    "typescript": "^3.4.5"
+    "@types/vscode": "^1.87.0",
+    "@types/mocha": "^10.0.6",
+    "@types/node": "18.x",
+    "@typescript-eslint/eslint-plugin": "^7.0.2",
+    "@typescript-eslint/parser": "^7.0.2",
+    "eslint": "^8.56.0",
+    "typescript": "^5.3.3",
+    "@vscode/test-cli": "^0.0.6",
+    "@vscode/test-electron": "^2.3.9"
+  },
+  "dependencies": {
+    "vscode-languageclient": "9.0.1"
   }
 }
