+++ vscode-ext/src/APIRegister.ts
