+++ vscode-ext/src/manager/Manager.ts
