+++ vscode-ext/src/manager/StateManager.ts
