+++ vscode-ext/src/manager/VerificationManager.ts
