+++ vscode-ext/src/utils.ts
