+++ vscode-ext/src/view/GutterInformationView.ts
