+++ vscode-ext/src/view/InlineGhostTextView.ts
