+++ vscode-ext/src/view/StatusBarView.ts
