+++ vscode-ext/src/view/ViewCollection.ts
